+++ Examples/UBLOX_Test/UBLOX_Test.ino
@@ -157,7 +157,7 @@ void loop()
 
 // Call back from within parser, for end of message
 // Process a complete message incoming from parser
-void processMessage(SEMP_PARSE_STATE *parse, uint8_t type)
+void processMessage(SEMP_PARSE_STATE *parse, uint16_t type)
 {
     static bool displayOnce = true;
     uint32_t offset;
